refactor(utils): split cli-path / log / pcm into dedicated subpath exports
Pulls the Node-facing CLI / logging / PCM helpers out of `@be-music/
utils` package root and exposes them as narrow subpath entry points
(`@be-music/utils/cli-path`, `/log`, `/pcm`). Browser bundles only
needed `core` / `path` / the workerize lazy-import escape hatch
before; the rest sat on the package-root re-export, which routed
every browser import through code that pulls in `node:fs/promises`,
`node:path`, etc., even though the demo never reaches those calls.

Changes:

- `packages/utils/src/cli-path.ts` (new) — `resolveCliPath` plus its
  Windows-drive / POSIX-absolute / lazy-`require` helpers, lifted
  verbatim from `index.ts`. The recent absolute-path fast-path fix
  moves with it.
- `packages/utils/src/index.ts` — re-exports `./cli-path.ts` so the
  package-root surface stays stable for any in-tree caller that
  hasn't migrated to the subpath yet (none after this commit, but
  cheap insurance).
- `packages/utils/package.json` — adds `./cli-path`, `./log`,
  `./pcm` to `exports`. (`./log` and `./pcm` already existed in
  source but only as package-root re-exports; this surfaces them as
  individually importable.)
- `packages/utils/tsdown.config.ts` — emits dist entries for the
  three new subpaths so the published package's `source` /
  `default` conditions both resolve.
- `tsconfig.typecheck.json` / `vitest.config.ts` /
  `packages/player-web-demo/vite.config.ts` — adds matching path
  aliases so type-check, vitest, and the demo all resolve the
  subpaths consistently.
- All consumers updated to import from the narrow subpath instead
  of the package root: TUI runtime, audio-renderer engine + CLIs,
  editor / parser / stringifier CLIs, the player engine's
  `core/engine.ts`. The tightening pulls `clamp` / `isAbortError` /
  `throwIfAborted` etc. from `@be-music/utils/core` and keeps the
  package root reserved for "no specific subpath cares" usage.

All 938 tests still pass; typecheck clean across utils / player /
player-tui / player-web / player-web-demo.
