+++ packages/audio-renderer/src/cli.ts
@@ -1,6 +1,6 @@
 import { resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { resolveCliPath } from '@be-music/utils';
+import { resolveCliPath } from '@be-music/utils/cli-path';
 import { renderChartFile } from './index.ts';
 
 interface CliArgs {
