+++ packages/player-tui/src/bga-video.ts
@@ -2,7 +2,7 @@ import { readFile } from 'node:fs/promises';
 import { basename } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { Worker } from 'node:worker_threads';
-import { createAbortError, isAbortError, throwIfAborted } from '@be-music/utils';
+import { createAbortError, isAbortError, throwIfAborted } from '@be-music/utils/core';
 
 const SUPPORTED_VIDEO_CODECS = new Set(['mpeg1video', 'h264', 'mjpeg']);
 // Keep chunks small so ff_decode_multi never allocates too many full-size frames at once.
